Calculate individual player scores

The Athletes Unlimited league plays matches of 3 sets, each to 25 points. To apply their scoring system to normal volleyball games, we need to decide what to do with sets 4 and 5 (if played). If you only want to count things that happened in the first 3 (or 4) sets, then you should remove data rows associated with sets 5 and/or 4 before calling this function. Otherwise, the set_5_weighting parameter allows the count of actions from set 5 to be scaled.

au_individual(
  x,
  set_5_weighting = 25/15,
  scoring = au_scoring(),
  infer_sets = FALSE,
  no_set_attacks = c("PR", "PP", "P2")
)

Arguments

x
datavolley or datavolleyplays: a datavolley object as returned by datavolley::dv_read(), or the plays component of one

set_5_weighting
numeric: the weighting that should be applied to actions from the fifth set of a match (if it went that long)

scoring
list: the scores to apply for each outcome. See au_scoring() for details

infer_sets
logical: set this to TRUE if the scout has not included all setting actions. Some scouts will not include setting actions unless the set is an error, or made by the non-designated setter. Using infer_sets = TRUE will insert a new setting action (made by the on-court setter) on attacks that don't have a corresponding set action scouted (excluding those in no_set_attacks)

no_set_attacks
character: a vector of attack codes for which sets are not expected (setter dumps, attacks on overpasses, second-ball attacks)

Value

A tibble with a breakdown by player of points in each category, along with

points : each player's total points scored. If set 5 results have been included, points scored in set 5 will be scaled by set_5_weighting

points_per_set: each player's total points divided by the number of sets they played

adjusted_points: each player's total points but adjusted for game length. The adjusted_points = points_per_set * 3 * nmatches * playing_time where nmatches is the average number of matches played by each team, and playing_time is (number of sets played by the player) / (number of sets played by their team)

Details

Most of the actions that are counted are self-explanatory: serve aces, attack kills, and serve, pass, attack, and set errors. The perhaps-less obvious ones:

A pass (serve reception) is counted if it is a perfect or positive pass

A set assist is a set that leads to an attack kill

Digs are counted on opposition attacks and block cover, if scouted, but not freeball passes. All dig grades except dig errors and digs directly back over the net are counted

Only block kills are counted

References

https://auprosports.com/volleyball/how-we-play-volleyball/

See also

Examples

x <- datavolley::dv_read(datavolley::dv_example_file())
au_individual(x)

#>          player_name player_id sets_played aces serve_errors serve_points
#> 1      NIKA MARKOVIČ       BR9           3    3            1           28
#> 2      ANITA SOBOČAN       MB9           3    4            1           40
#> 3    ANA MARIJA VOVK       BR6           3    2            1           16
#> 4        DARJA ERŽEN       MB7           3    3            2           20
#> 5    ŠPELA BEDNJANIČ       BR1           3    0            0            0
#> 6    KATJA MIHALINEC      BR12           3    2            2            8
#> 7     ZALA ZBIČAJNIK       BR4           3    0            1           -8
#> 8       KATJA MIHEVC      BR10           3    0            1           -8
#> 9      KAJA KEGLEVIČ       MB1           3    2            1           16
#> 10      KATJA ČEPLAK      MB10           3    0            2          -16
#> 11 ANJA HRIBERNIK 97      BR15           3    1            0           12
#> 12    KARMINA SUŠNIK      MB11           3    0            0            0
#> 13       MIA LEDINEK       MB2           3    1            2           -4
#> 14     LAURA MUNIŠIČ       MB4           2    0            0            0
#> 15    MAJA ZEMLJARIČ       MB6           2    0            0            0
#> 16      NIKA MUHOVIC       BR8           2    0            1           -8
#> 17        ELA PINTAR       MB8           3    0            2          -16
#> 18   SAMANTA RITUPER       BR5           1    0            1           -8
#>    passes pass_errors pass_points assists set_errors set_points kills
#> 1       0           1         -12       0          0          0    18
#> 2       6           0          12       0          0          0     7
#> 3       3           1          -6       0          0          0     3
#> 4       0           0           0       0          1        -12    11
#> 5       2           2         -20       1          0          1     0
#> 6       4           6         -64       0          0          0    10
#> 7       0           0           0       8          1         -4     2
#> 8       0           0           0       0          0          0     2
#> 9       0           0           0      11          1         -1     0
#> 10      7           1           2       0          0          0     3
#> 11      0           0           0       0          0          0     3
#> 12     13           7         -58       2          0          2     0
#> 13      0           0           0       0          0          0     0
#> 14      0           0           0       0          0          0     0
#> 15      0           0           0       0          0          0     0
#> 16      0           0           0       0          0          0     0
#> 17      1           0           2       0          0          0     2
#> 18      0           0           0       0          0          0     0
#>    attack_errors attack_points blocks block_points digs dig_points points
#> 1              2           120      2           24    6         30    190
#> 2              3            20      0            0    8         40    112
#> 3              0            24      0            0   11         55     89
#> 4              3            52      1           12    2         10     82
#> 5              0             0      0            0   16         80     61
#> 6              2            56      1           12    8         40     52
#> 7              0            16      1           12    7         35     51
#> 8              0            16      2           24    2         10     42
#> 9              0             0      0            0    4         20     35
#> 10             1            12      0            0    7         35     33
#> 11             1            12      0            0    1          5     29
#> 12             0             0      0            0   14         70     14
#> 13             0             0      1           12    0          0      8
#> 14             0             0      0            0    0          0      0
#> 15             0             0      0            0    0          0      0
#> 16             0             0      0            0    0          0     -8
#> 17             3           -20      1           12    1          5    -17
#> 18             0             0      0            0    0          0     -8
#>    adjusted_points points_per_set
#> 1              190      63.333333
#> 2              112      37.333333
#> 3               89      29.666667
#> 4               82      27.333333
#> 5               61      20.333333
#> 6               52      17.333333
#> 7               51      17.000000
#> 8               42      14.000000
#> 9               35      11.666667
#> 10              33      11.000000
#> 11              29       9.666667
#> 12              14       4.666667
#> 13               8       2.666667
#> 14               0       0.000000
#> 15               0       0.000000
#> 16              -8      -4.000000
#> 17             -17      -5.666667
#> 18              -8      -8.000000